feat(minimap): hover tooltip shows repo, branch, and last-active duration
Closes #870. Replaces the opaque `title={id}` on minimap tile rects and ghost markers with the same identity projection the workspace switcher uses (group · label · #suffix) plus a 'last active 5m ago' line driven by the existing formatTimeAgo. The formatter moves from SearchPanel.tsx into staleness.ts (third caller would be the elephant in the room; pull it out at the second).
